feat: Adding support for extraContainers and StatefuleSet deployment
The purpose of these modifications is to add support for deploying the registry container as a StatefuleSet STS instead of a Deployment. This need several modifications on the templates, because the STS handles volume differently than the Deployment. A source of instpiration for this modification is the Grafana Helm Chart https://github.com/grafana/helm-charts/tree/main/charts/grafana. Adding support for extraContainers to be able to add sidecar containers to the registry container.

Signed-off-by: Nicolas-Peiffer <102670102+Nicolas-Peiffer@users.noreply.github.com>